First time
Unfortunately, I just started working on the GUI, and I spent roughly one hour on this code; I learned about ActionListener, which is an interface for using live events. I wanted to use this to create a button program. Now i want to add if statements and give a little pop up quiz on wednesday!